Fix central + others
